+++ graphics/custom/render_present.h
@@ -12,9 +12,7 @@
 namespace Verse::Graphics
 {
     void present(Config &c, SDL_Renderer *renderer, SDL_Window *window,
-                 SDL_Texture *render_target, SDL_Texture *palette_tex,
-                 int pid, int width, int height);
+                 SDL_Texture *render_target, SDL_Texture *palette_tex, int pid);
 
     void handlePaletteTransition(Config &c);
-    void finishPaletteTransition(Config &c);
 }