+++ graphics/gui.cpp
@@ -11,12 +11,16 @@
 
 using namespace Verse;
 
-void Gui::update(float delta) {
+void Gui::update(float delta, Config &c) {
     ImGuiIO& imgui_io = ImGui::GetIO();
     
     imgui_io.DeltaTime = delta;
     
-    imgui_io.MousePos = ImVec2(static_cast<float>(Input::mouse().x), static_cast<float>(Input::mouse().y));
+    Vec2 mouse_pos = Vec2(Input::mouse().x, Input::mouse().y);
+    if (c.use_shaders)
+        mouse_pos -= c.window_padding;
+    
+    imgui_io.MousePos = ImVec2(mouse_pos.x, mouse_pos.y);
     imgui_io.MouseDown[0] = Input::down(SDL_BUTTON_LEFT);
     imgui_io.MouseDown[1] = Input::down(SDL_BUTTON_RIGHT);
     imgui_io.MouseWheel = static_cast<float>(Input::mouseWheel());