+++ graphics/gui.h
@@ -17,7 +17,7 @@
 
 namespace Verse::Gui
 {
-    void update(float delta);
+    void update(float delta, Config &c);
     void prerender(Scene &scene, Config &c, ui16 &fps);
     void render();
 }